Move main into separate file
