+++ action.yml
@@ -37,4 +37,4 @@ outputs:
 
 runs:
   using: "node20"
-  main: "dist/index.js"
+  main: "dist/main.js"
