+++ package.json
@@ -4,11 +4,11 @@
   "description": "A GitHub Action for downloading and installing LLVM and Clang binaries.",
   "author": "Kyle Mayes <kyle@mayeses.com>",
   "license": "Apache-2.0",
-  "main": "dist/index.js",
+  "main": "dist/main.js",
   "scripts": {
     "format": "prettier --write *.ts",
     "generate": "ts-node generate.ts",
-    "build": "parcel build index.ts",
+    "build": "parcel build main.ts",
     "test": "ts-node test.ts test"
   },
   "dependencies": {
